feat(spx-gui): add Map Editor's SpriteBasicConfig
